api-py/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tabula-py==2.9.0
pandas==2.1.3
